chore: удален компонент Button
Компонент Button2 переименован в Button
