+++ src/components/constructor-link/constructor-link.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames/bind';
 
-import { Button } from 'components/ui/button2';
+import { Button } from 'components/ui/button';
 import { Icon } from 'components/ui/icon';
 
 import type { VFC } from 'react';
