+++ src/components/festival-event-card/festival-event-card.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames/bind';
 
-import { Button } from 'components/ui/button2';
+import { Button } from 'components/ui/button';
 import { Icon } from 'components/ui/icon';
 import { CreditsList } from 'components/credits-list';
