+++ src/components/for-press-hero/for-press-hero-description/for-press-hero-description.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import cn from 'classnames/bind';
 
-import { Button } from 'components/ui/button2';
+import { Button } from 'components/ui/button';
 import { Icon } from 'components/ui/icon';
 import { Url } from 'shared/types/common';
