+++ src/components/homepage-video-archive-section/homepage-video-archive-section.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import classNames from 'classnames/bind';
 
-import { Button } from 'components/ui/button2';
+import { Button } from 'components/ui/button';
 import { Icon } from 'components/ui/icon';
 
 import type { FC } from 'react';
