+++ src/components/teaser-card/teaser-card.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames/bind';
 import Image from 'next/image';
 import Link from 'next/link';
 
-import { Button } from 'components/ui/button2';
+import { Button } from 'components/ui/button';
 import { Icon } from 'components/ui/icon';
 
 import type { FC } from 'react';
