--- src/components/ui/button2/button.module.css
@@ -1,177 +1,0 @@
-$border-width: 1px;
-
-.common {
-  position: relative;
-  display: inline-flex;
-  box-sizing: border-box;
-  align-items: center;
-  justify-content: space-between;
-  border: $border-width solid;
-  background: 0;
-  color: inherit;
-  cursor: pointer;
-  isolation: isolate;
-  text-decoration: none;
-
-  &:hover {
-    border-color: transparent;
-    color: var(--white);
-
-    &::before {
-      position: absolute;
-      z-index: -1;
-      top: -$border-width;
-      left: -$border-width;
-      display: block;
-      width: calc(100% + $border-width);
-      height: calc(100% + $border-width);
-      animation: .3s background-slide ease;
-      background-color: var(--coal);
-      content: '';
-      transform-origin: left;
-    }
-  }
-}
-
-.icon {
-  display: block;
-  width: 20px;
-  height: 20px;
-  flex-shrink: 0;
-  font-size: 0;
-  line-height: 0;
-}
-
-.icon-left {
-  composes: icon;
-}
-
-.icon-right {
-  composes: icon;
-  order: 1;
-}
-
-.text {
-  @mixin text;
-}
-
-.xs {
-  composes: common;
-
-  .text {
-    font-size: 14px;
-    line-height: 16px;
-  }
-}
-
-.s {
-  composes: common;
-  padding: 6px;
-
-  .icon-left:not(:only-child) {
-    margin-right: 8px;
-  }
-
-  .icon-right:not(:only-child) {
-    margin-left: 8px;
-  }
-
-  .text {
-    font-size: 16px;
-    line-height: 20px;
-  }
-}
-
-.m {
-  composes: common;
-  padding: 8px;
-
-  .icon-left:not(:only-child) {
-    margin-right: 8px;
-  }
-
-  .icon-right:not(:only-child) {
-    margin-left: 8px;
-  }
-
-  .text {
-    font-size: 20px;
-    line-height: 28px;
-  }
-}
-
-.l {
-  composes: common;
-  padding: 16px;
-
-  .icon-left:not(:only-child) {
-    margin-right: 16px;
-  }
-
-  .icon-right:not(:only-child) {
-    margin-left: 16px;
-  }
-
-  .text {
-    font-size: 16px;
-    line-height: 22px;
-  }
-}
-
-.common.border-none {
-  border-color: transparent;
-}
-
-.common.border-full {
-  border-color: var(--coal);
-}
-
-.common.border-right-bottom-left {
-  border-color: var(--coal);
-  border-top-color: transparent;
-}
-
-.common.border-bottom-left {
-  border-top-color: transparent;
-  border-right-color: transparent;
-}
-
-.common.border-right-bottom {
-  border-top-color: transparent;
-  border-left-color: transparent;
-}
-
-.common.border-top-left {
-  border-right-color: transparent;
-  border-bottom-color: transparent;
-}
-
-.common.border-top {
-  border-color: transparent;
-  border-top: 1px solid var(--coal);
-}
-
-.common.full-width {
-  width: 100%;
-}
-
-.common.upper-case {
-  .text {
-    text-transform: uppercase;
-  }
-}
-
-.common.disabled {
-  color: var(--gray);
-  pointer-events: none;
-}
-
-@keyframes background-slide {
-  from {
-    transform: scaleX(0);
-  }
-
-  to {
-    transform: scaleX(1);
-  }
-}
