--- src/components/ui/button2/index.ts
@@ -1,1 +1,0 @@
-export * from './button';
