+++ src/pages/library/search-no-result/index.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/router';
 import * as breakpoints from 'shared/breakpoints.js';
 import { AppLayout } from 'components/app-layout';
 import { SEO } from 'components/seo';
-import { Button } from 'components/ui/button2';
+import { Button } from 'components/ui/button';
 import { Icon } from 'components/ui/icon';
 import { useMediaQuery } from 'shared/hooks/use-media-query';
 import LibraryForm from 'components/library-form/library-form';
